requirements.txt
python-dotenv>=1.0.0
pandas>=2.3.0
numpy>=1.26.0
homeharvest>=0.8.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0
tqdm>=4.66.0
oauth2client>=4.1.3
gspread>=6.0.0
gspread-dataframe>=3.3.0
requests>=2.32.0
sib-api-v3-sdk>=7.6.0
